Changed data type to double to allow for decimals

+++ BasicCalculator/Program.cs
@@ -11,9 +11,9 @@ namespace BasicCalculator
         static void Main(string[] args)
         {
             Console.Write("Enter a number: ");
-            int num1 = Convert.ToInt32(Console.ReadLine());
+            double num1 = Convert.ToDouble(Console.ReadLine());
             Console.Write("Enter another number: ");
-            int num2 = Convert.ToInt32(Console.ReadLine());
+            double num2 = Convert.ToDouble(Console.ReadLine());
 
             Console.WriteLine("{0} + {1} = {2}", num1, num2, num1 + num2);
             Console.ReadLine();
